+++ src/containers/AddEmail.js
@@ -1,5 +1,5 @@
 import { connect } from "react-redux";
-import Email from "components/Email";
+import AddEmail from "components/AddEmail";
 import * as actions from "actions/Email";
 import i18n from "../login/translation/InjectIntl_HOC_factory";
 import { history } from "components/SignupMain";
@@ -46,6 +46,6 @@ const mapDispatchToProps = (dispatch, props) => {
   };
 };
 
-const EmailContainer = connect(mapStateToProps, mapDispatchToProps)(Email);
+const AddEmailContainer = connect(mapStateToProps, mapDispatchToProps)(AddEmail);
 
-export default i18n(EmailContainer);
+export default i18n(AddEmailContainer);
